fix: new disclaimer for net apy tooltip
